# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.ImageMetadataResponse:
    properties:
      height:
        type: integer
      mime_type:
        type: string
      size:
        type: integer
      width:
        type: integer
    type: object
  dto.ListImagesResponse:
    properties:
      images:
        items:
          $ref: '#/definitions/image.Image'
        type: array
      total:
        type: integer
    type: object
  dto.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.RegisterRequest:
    properties:
      password:
        minLength: 8
        type: string
      username:
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  dto.TransformResponse:
    properties:
      height:
        type: integer
      id:
        type: string
      mime_type:
        type: string
      size:
        type: integer
      variant_key:
        type: string
      width:
        type: integer
    type: object
  dto.UploadResponse:
    properties:
      id:
        type: string
      metadata:
        $ref: '#/definitions/dto.ImageMetadataResponse'
      original_url:
        type: string
    type: object
  dto.UserResponse:
    properties:
      id:
        type: string
      username:
        type: string
    type: object
  image.CropSpec:
    properties:
      height:
        maximum: 8000
        minimum: 1
        type: integer
      width:
        maximum: 8000
        minimum: 1
        type: integer
      x:
        minimum: 0
        type: integer
      "y":
        minimum: 0
        type: integer
    required:
    - height
    - width
    type: object
  image.FilterSpec:
    properties:
      blur:
        description: sigma
        type: integer
      grayscale:
        type: boolean
      sepia:
        type: boolean
    type: object
  image.Image:
    properties:
      createdAt:
        type: string
      filename:
        type: string
      height:
        type: integer
      id:
        type: string
      mimeType:
        type: string
      originalKey:
        type: string
      ownerID:
        type: string
      size:
        format: int64
        type: integer
      variants:
        items:
          $ref: '#/definitions/image.Variant'
        type: array
      width:
        type: integer
    type: object
  image.ResizeSpec:
    properties:
      height:
        maximum: 8000
        minimum: 1
        type: integer
      width:
        maximum: 8000
        minimum: 1
        type: integer
    type: object
  image.TransformationSpec:
    properties:
      crop:
        $ref: '#/definitions/image.CropSpec'
      filters:
        $ref: '#/definitions/image.FilterSpec'
      flip:
        type: boolean
      format:
        enum:
        - jpeg
        - png
        - webp
        - gif
        type: string
      mirror:
        type: boolean
      quality:
        maximum: 100
        minimum: 1
        type: integer
      resize:
        $ref: '#/definitions/image.ResizeSpec'
      rotate:
        enum:
        - 0
        - 90
        - 180
        - 270
        type: integer
      watermark:
        $ref: '#/definitions/image.WatermarkSpec'
    type: object
  image.Variant:
    properties:
      createdAt:
        type: string
      height:
        type: integer
      id:
        type: string
      mimeType:
        type: string
      size:
        description: File size in bytes
        format: int64
        type: integer
      specHash:
        description: Hash of the transformation spec
        type: string
      variantKey:
        description: Storage key (path)
        type: string
      width:
        type: integer
    type: object
  image.WatermarkSpec:
    properties:
      gravity:
        description: center, north, southeast, etc.
        type: string
      image_id:
        type: string
      opacity:
        description: 0.0-1.0
        type: number
      text:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: High-performance image processing service with synchronous and asynchronous
    transformations.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Image Processing Service API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "401":
          description: Invalid credentials
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: User login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with username and password
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: User already exists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - auth
  /images:
    get:
      description: List all images owned by the user with pagination
      parameters:
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      - default: 10
        description: Limit for pagination
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of images
          schema:
            $ref: '#/definitions/dto.ListImagesResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List user images
      tags:
      - images
    post:
      consumes:
      - multipart/form-data
      description: Upload a new image for processing. Requires multipart/form-data.
      parameters:
      - description: Image file to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Image uploaded successfully
          schema:
            $ref: '#/definitions/dto.UploadResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Upload an image
      tags:
      - images
  /images/{id}:
    get:
      description: Fetch metadata and variants for a specific image
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Image details
          schema:
            $ref: '#/definitions/image.Image'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Image not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get image details
      tags:
      - images
  /images/{id}/transform:
    post:
      consumes:
      - application/json
      description: Apply transformations to an image. Use sync=true for immediate
        response.
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: string
      - description: Perform transformation synchronously
        in: query
        name: sync
        type: boolean
      - description: Transformation metadata
        in: body
        name: spec
        required: true
        schema:
          $ref: '#/definitions/image.TransformationSpec'
      produces:
      - application/json
      responses:
        "200":
          description: Transformation result (sync)
          schema:
            $ref: '#/definitions/dto.TransformResponse'
        "202":
          description: Transformation accepted (async)
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Transform an image
      tags:
      - images
  /me:
    get:
      description: Get details of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User profile details
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"